requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
automated-interpretability==0.0.13
babe==0.0.7
babel==2.17.0
beartype==0.14.1
beautifulsoup4==4.13.4
better-abc==0.0.3
bidict==0.23.1
bitsandbytes==0.46.1
bleach==6.2.0
blobfile==2.1.1
boostedblob==0.15.6
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
config2py==0.1.37
contourpy==1.3.2
cut-cross-entropy==25.1.1
cycler==0.12.1
datasets
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.34.0
dill==0.3.8
distro==1.9.0
docstring-parser==0.16
dol==0.3.19
einops==0.8.1
evaluate==0.4.5
exceptiongroup==1.3.0
executing==2.2.0
fancy-einsum==0.0.3
fastjsonschema==2.21.1
filelock==3.18.0
fire
fonttools==4.59.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2024.6.1
gitdb==4.0.12
gitpython==3.1.44
gprof2dot==2025.4.14
graze==0.1.29
h11==0.16.0
hf-transfer==0.1.9
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.33.2
i2==0.1.48
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
iniconfig==2.1.0
ipykernel==6.30.0
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
jaxtyping==0.3.2
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.8
lark==1.2.2
lxml==4.9.4
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nnsight==0.4.10
notebook==7.4.4
notebook-shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai==1.97.1
orjson==3.11.1
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
peft==0.15.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==5.24.1
plotly-express==0.4.1
pluggy==1.6.0
prometheus-client==0.22.1
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py2store==0.1.20
pyarrow==20.0.0
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-profiling==1.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-json-logger==3.3.0
python-socketio==5.13.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.0.0
rpds-py==0.26.0
sae-lens
safetensors==0.4.5
scikit-learn==1.7.1
scipy==1.15.3
send2trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.32.0
setuptools==80.9.0
shellingham==1.5.4
shtab==1.7.2
simple-parsing==0.1.7
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.5
steering-vectors==0.12.2
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformer-lens==2.16.1
transformers==4.51.3
transformers-stream-generator==0.0.5
triton==3.3.0
trl==0.19.0
typeguard==4.4.4
typer==0.12.5
types-python-dateutil==2.9.0.20250708
typing-extensions==4.14.0
typing-inspection==0.4.1
tyro==0.9.26
tzdata==2025.2
unsloth==2025.6.12
unsloth-zoo==2025.6.8
uri-template==1.3.0
urllib3==2.5.0
uvloop==0.21.0
wadler-lindig==0.1.7
wandb==0.21.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
widgetsnbextension==4.0.14
wsproto==1.2.0
xformers==0.0.30
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.22.0
goodfire
claude-agent-sdk
fastapi
uvicorn
